Relacion OneToOne unidireccional y modificacion de botones

+++ tp4_VelizGuadalupe/src/main/java/ar/edu/unju/fi/service/imp/MateriaServiceImp.java
@@ -2,24 +2,35 @@ package ar.edu.unju.fi.service.imp;
 
 import java.util.List;
 import java.util.stream.Collectors;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import ar.edu.unju.fi.dto.MateriaDTO;
 import ar.edu.unju.fi.mapper.MateriaMapper;
+import ar.edu.unju.fi.model.Alumno;
+import ar.edu.unju.fi.model.Carrera;
 import ar.edu.unju.fi.model.Materia;
+import ar.edu.unju.fi.repository.AlumnoRepository;
+import ar.edu.unju.fi.repository.CarreraRepository;
 import ar.edu.unju.fi.repository.MateriaRepository;
 import ar.edu.unju.fi.service.MateriaService;
+
 
 @Service
 public class MateriaServiceImp implements MateriaService {
 
 	@Autowired
 	private MateriaRepository materiaRepository;
+	
+    @Autowired
+    private AlumnoRepository alumnoRepository;
+
+    @Autowired
+    private CarreraRepository carreraRepository;
 
 	@Autowired
 	private MateriaMapper materiaMapper;
+	
+	
 
 	@Override
 	public MateriaDTO getMateria(Long id) {
@@ -34,16 +45,16 @@ public class MateriaServiceImp implements MateriaService {
 	@Override
 	public List<MateriaDTO> getMaterias() {
 		List<Materia> materias = materiaRepository.findAll();
-		return materias.stream().map(materia -> {
-			return materiaMapper.materiaToMateriaDTO(materia);
-		}).collect(Collectors.toList());
+		return materias.stream()
+				.map(materia -> {
+					return materiaMapper.materiaToMateriaDTO(materia);
+				}).collect(Collectors.toList());
 	}
 
 	@Override
 	public void saveMateria(MateriaDTO materiaDTO) {
 		Materia materia = materiaMapper.materiaDTOToMateria(materiaDTO);
 		materiaRepository.save(materia);
-
 	}
 
 	@Override
@@ -54,8 +65,23 @@ public class MateriaServiceImp implements MateriaService {
 
 	@Override
 	public void deleteMateria(Long id) {
-		materiaRepository.deleteById(id);
-
+		Materia materia = materiaRepository.findById(id).orElse(null);
+		
+		if (materia != null) {
+			
+			for (Alumno alumno : materia.getAlumnos()) {
+				alumno.getMaterias().remove(materia);
+				alumnoRepository.save(alumno);
+			}
+			
+			Carrera carrera = materia.getCarrera();
+			if (carrera != null) {
+				carrera.getMaterias().remove(materia);
+				carreraRepository.save(carrera);
+			}
+			
+			materiaRepository.deleteById(id);
+		}		
 	}
 
 }
